# docs/openapi(2) (1).yaml
openapi: 3.0.3
info:
  title: Onidel Cloud API - OpenAPI 3.0
  description: |-
    Welcome to the **Onidel Cloud API** documentation. This API allows developers to seamlessly integrate with the Onidel Cloud platform, enabling them to manage cloud resources efficiently and securely. With our API, you can automate processes, retrieve detailed account information, and manage virtual machines (VPS), networking, storage, and more.
  termsOfService: https://docs.onidel.com/policies/terms-of-service
  contact:
    email: support@onidel.com
  version: 1.2.0
servers:
  - url: https://api.cloud.onidel.com
...
paths:
...
  /measured-boot-images:
    get:
      tags:
        - Measured Boot Image
      summary: List measured boot images
      description: List the team's measured direct boot images (UKIs).
      operationId: getMeasuredBootImages
      security:
        - api_key: []
      parameters:
        - name: team_id
          in: query
          description: Team ID. Default team will be used if not provided.
          required: false
          explode: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    $ref: '#/components/schemas/MeasuredBootImages'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags:
        - Measured Boot Image
      summary: Upload a measured boot image
      description: |-
        Upload a UKI (unified kernel image bundling kernel, initrd and cmdline)
        for SEV-SNP measured direct boot. Subject to the maximum file size limit.
      operationId: uploadMeasuredBootImage
      security:
        - api_key: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The UKI file.
                team_id:
                  type: string
                  format: uuid
                  description: Team ID. Default team will be used if not provided.
                  example: 96616865-475e-4191-aba5-f89f0a4b71fe
                description:
                  type: string
                  description: Optional description.
                  example: debian-12 measured boot
      responses:
        '201':
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasuredBootImage'
        '400':
          description: Bad request (no file or file too big)
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /measured-boot-images/{image_id}:
    delete:
      tags:
        - Measured Boot Image
      summary: Delete a measured boot image
      description: Delete a measured boot image. It must not be attached to any instance.
      operationId: deleteMeasuredBootImage
      security:
        - api_key: []
      parameters:
        - name: image_id
          in: path
          description: Measured boot image UUID
          required: true
          explode: true
          schema:
            type: string
            format: uuid
        - name: team_id
          in: query
          description: Team ID. Default team will be used if not provided.
          required: false
          explode: true
          schema:
            type: string
      responses:
        '204':
          description: Successful operation
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: Conflict (image still attached to an instance)
...
  /vm/{vm_id}/attach-measured-boot:
    post:
      tags:
        - Virtual Machine
      summary: Attach a measured boot image
      description: Attach a measured boot image (UKI) to the VM and power-cycle to apply it. SEV-SNP must already be enabled.
      operationId: attachMeasuredBoot
      security:
        - api_key: []
      parameters:
        - name: vm_id
          in: path
          description: VM UUID
          required: true
          explode: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - uki_image_id
              properties:
                team_id:
                  description: Team ID. Default team will be used if not provided.
                  type: string
                  format: uuid
                  example: 96616865-475e-4191-aba5-f89f0a4b71fe
                uki_image_id:
                  description: Measured boot image (UKI) ID to attach.
                  type: string
                  format: uuid
                  example: 7c1f2a90-4b3d-4e21-9a8c-1d2e3f4a5b6c
      responses:
        '202':
          description: Attach started
          content:
            application/json:
              schema:
                type: object
                properties:
                  audit_log_id:
                    type: string
                    format: uuid
                    example: 0f9c2d1a-3b4e-4c5d-8a6f-7b8c9d0e1f2a
        '400':
          description: Bad request (missing image or SEV-SNP not enabled)
        '401':
          description: Unauthorized
        '404':
          description: Not found
  /vm/{vm_id}/detach-measured-boot:
    post:
      tags:
        - Virtual Machine
      summary: Detach the measured boot image
      description: Clear the VM's measured boot image and power-cycle to apply.
      operationId: detachMeasuredBoot
      security:
        - api_key: []
      parameters:
        - name: vm_id
          in: path
          description: VM UUID
          required: true
          explode: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                team_id:
                  description: Team ID. Default team will be used if not provided.
                  type: string
                  format: uuid
                  example: 96616865-475e-4191-aba5-f89f0a4b71fe
      responses:
        '202':
          description: Action accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  audit_log_id:
                    type: string
                    format: uuid
                    example: 0f9c2d1a-3b4e-4c5d-8a6f-7b8c9d0e1f2a
        '401':
          description: Unauthorized
        '404':
          description: Not found
...
components:
  schemas:
...
    MeasuredBootImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 7c1f2a90-4b3d-4e21-9a8c-1d2e3f4a5b6c
        filename:
          type: string
          example: debian-12.uki.efi
        description:
          type: string
          example: debian-12 measured boot
        size:
          type: integer
          format: int64
          description: Size in bytes.
          example: 52428800
        created:
          type: string
          format: datetime
          example: 2025-01-05T13:00:20.010960Z
    MeasuredBootImages:
      type: array
      items:
        $ref: '#/components/schemas/MeasuredBootImage'
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer